refactor a lot of attribute code
