+++ src/generators/dom/visitors/Element/Element.js
@@ -1,11 +1,28 @@
 import deindent from '../../../../utils/deindent.js';
 import visit from '../../visit.js';
-import addElementAttributes from './attributes/addElementAttributes.js';
 import visitComponent from '../Component/Component.js';
 import visitWindow from './meta/Window.js';
+import visitAttribute from './Attribute.js';
+import visitEventHandler from './EventHandler.js';
+import visitBinding from './Binding.js';
+import visitRef from './Ref.js';
 
 const meta = {
 	':Window': visitWindow
+};
+
+const order = {
+	Attribute: 1,
+	EventHandler: 2,
+	Binding: 3,
+	Ref: 4
+};
+
+const visitors = {
+	Attribute: visitAttribute,
+	EventHandler: visitEventHandler,
+	Binding: visitBinding,
+	Ref: visitRef
 };
 
 export default function visitElement ( generator, block, state, node ) {
@@ -22,10 +39,12 @@ export default function visitElement ( generator, block, state, node ) {
 	const childState = Object.assign( {}, state, {
 		isTopLevel: false,
 		parentNode: name,
-		namespace: node.name === 'svg' ? 'http://www.w3.org/2000/svg' : state.namespace
+		namespace: node.name === 'svg' ? 'http://www.w3.org/2000/svg' : state.namespace,
+		allUsedContexts: []
 	});
 
 	block.builders.create.addLine( `var ${name} = ${getRenderStatement( generator, childState.namespace, node.name )};` );
+	block.mount( name, state.parentNode );
 
 	if ( !state.parentNode ) {
 		block.builders.detach.addLine( `${generator.helper( 'detachNode' )}( ${name} );` );
@@ -36,14 +55,25 @@ export default function visitElement ( generator, block, state, node ) {
 		block.builders.create.addLine( `${generator.helper( 'setAttribute' )}( ${name}, '${generator.cssId}', '' );` );
 	}
 
-	const local = {
-		allUsedContexts: []
-	};
+	node.attributes
+		.sort( ( a, b ) => order[ a.type ] - order[ b.type ] )
+		.forEach( attribute => {
+			visitors[ attribute.type ]( generator, block, childState, node, attribute );
+		});
 
-	addElementAttributes( generator, block, childState, node, local );
+	// special case – bound <option> without a value attribute
+	if ( node.name === 'option' && !node.attributes.find( attribute => attribute.type === 'Attribute' && attribute.name === 'value' ) ) { 	// TODO check it's bound
+		const statement = `${name}.__value = ${name}.textContent;`;
+		block.builders.update.addLine( statement );
+		node.initialUpdate = statement;
+	}
 
-	if ( local.allUsedContexts.length ) {
-		const initialProps = local.allUsedContexts.map( contextName => {
+	if ( node.initialUpdate ) {
+		block.builders.create.addBlock( node.initialUpdate );
+	}
+
+	if ( childState.allUsedContexts.length ) {
+		const initialProps = childState.allUsedContexts.map( contextName => {
 			if ( contextName === 'root' ) return `root: root`;
 
 			const listName = block.listNames.get( contextName );
@@ -52,7 +82,7 @@ export default function visitElement ( generator, block, state, node ) {
 			return `${listName}: ${listName},\n${indexName}: ${indexName}`;
 		}).join( ',\n' );
 
-		const updates = local.allUsedContexts.map( contextName => {
+		const updates = childState.allUsedContexts.map( contextName => {
 			if ( contextName === 'root' ) return `${name}.__svelte.root = root;`;
 
 			const listName = block.listNames.get( contextName );
@@ -70,30 +100,17 @@ export default function visitElement ( generator, block, state, node ) {
 		block.builders.update.addBlock( updates );
 	}
 
-	// special case – bound <option> without a value attribute
-	if ( node.name === 'option' && !node.attributes.find( attribute => attribute.type === 'Attribute' && attribute.name === 'value' ) ) { 	// TODO check it's bound
-		const statement = `${name}.__value = ${name}.textContent;`;
-		block.builders.update.addLine( statement );
-		node.initialUpdate = statement;
-	}
-
-	block.createMountStatement( name, state.parentNode );
-
 	node.children.forEach( child => {
 		visit( generator, block, childState, child );
 	});
-
-	if ( node.initialUpdate ) {
-		block.builders.create.addBlock( node.initialUpdate );
-	}
 }
 
 function getRenderStatement ( generator, namespace, name ) {
-	if ( namespace ) {
-		if ( namespace === 'http://www.w3.org/2000/svg' ) {
-			return `${generator.helper( 'createSvgElement' )}( '${name}' )`;
-		}
+	if ( namespace === 'http://www.w3.org/2000/svg' ) {
+		return `${generator.helper( 'createSvgElement' )}( '${name}' )`;
+	}
 
+	if ( namespace ) {
 		return `document.createElementNS( '${namespace}', '${name}' )`;
 	}
