+++ src/generators/dom/visitors/Element/getStaticAttributeValue.js
