+++ src/generators/dom/visitors/Element/lookup.js
